+++ packages/app-builder/src/components/ui/header/PageSelector.tsx
@@ -5,11 +5,11 @@ import { usePageId } from '@/components/hooks/use-page-id'
 import Delete from '@/components/icons/delete'
 import Plus from '@/components/icons/plus'
 import { cx } from '@linaria/core'
-import { styled } from '@linaria/react'
-import { Button, elFlex, elFlex1, elFlexAlignCenter, elFlexJustifyStart, elM2, Select } from '@reapit/elements'
+import { elFlex, elFlex1, elFlexAlignCenter, elFlexJustifyStart, elM2 } from '@reapit/elements'
 import React from 'react'
 import slugify from 'slugify'
 import { emptyState } from '../../hooks/apps/emptyState'
+import { AppBuilderIconButton, AppBuilderSelect } from '../components'
 
 export const newPage = (name: string) => {
   const page = {
@@ -19,40 +19,6 @@ export const newPage = (name: string) => {
   }
   return page
 }
-
-const AppBuilderSelect = styled(Select)`
-  border: 1px solid #e3e3e3;
-  border-radius: 4px;
-  max-width: 297px;
-  margin-right: 4px;
-
-  &:focus {
-    border-bottom-color: #e3e3e3;
-  }
-`
-
-export const AppBuilderIconButton = styled(Button)`
-  width: 32px;
-  height: 32px;
-  padding: 0;
-  margin: 0;
-  margin-left: 4px;
-  margin-right: 4px;
-  border: none;
-  border-radius: 4px;
-
-  svg {
-    fill: white;
-  }
-
-  background: ${({ intent }) => {
-    if (intent === 'primary') {
-      return '#23A4DE'
-    }
-
-    return '#f2f2f2'
-  }};
-`
 
 export const PageSelector = ({ pageId, onChange }: { pageId?: string; onChange: (id: string) => void }) => {
   const { appId, setPageId } = usePageId()
